% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\docType{data}
\name{Constants}
\alias{Constants}
\alias{suppModels}
\alias{probModels}
\alias{contModels}
\title{Package Constants}
\format{A list of models that are supported. These represent the internal codes used by this package.}
\usage{
suppModels, probModels, contModels
}
\description{
Global constants that define the bounds of this packages functionality
}
\section{Constants}{

\itemize{
  \item \code{suppModels}: supported models---the model types (packages) this package currently supports.
  \item \code{probModels}: probabilistic models---models types that will return probabilities when created on categorical data.
  \item \code{contModels}: continuous models---models that can be used (natively) with continuous variables. Of course other
    classification algorithms, for example nearest neighbour, can be run on an index variable then continuous data can then be
    \code{\link[=impute]{imputed}} from this.
}
}

\examples{
\dontrun{
suppModels <- c('randomForest','rfsrc','fnn.FNN','fnn.class','kknn','gbm')  # Which packages (model classes) are supported by NPEL.Classification
probModels <- c('randomForest','rfsrc','fnn.FNN','fnn.class','kknn','gbm')  # Which packages generate probabilities when passed categorical data
contModels <- c('randomForest','rfsrc','kknn','gbm')                               # Which packages are able to handle continuous data
}
}
