% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\docType{data}
\name{ecoGroup}
\alias{ecoGroup}
\title{An example of a transformation 'function'}
\format{A matrix list with 4 rows (scenarios) and 3 columns (possible data types for that scenario)
\describe{
  \item{\sQuote{transform}}{a transformation vector converting input factors (Ecotypes) to output factors (groups). Functionally
     it is a lookup table that maps input classes to output classes. See examples for how to use it.}
  \item{\sQuote{labels}}{names for the output groups}
  \item{\sQuote{colours}}{suggested plotting colours for the output groups}
}}
\usage{
ecoGroup[['which.grouping','type.of.data']]
}
\arguments{
\item{which.grouping}{one of \code{c('identity','domSpecies','domGroup','maxGranularity'), where: \sQuote{identity} does no 
transformation, \sQuote{domSpecies} groups by the dominant species, \sQuote{domGroup} which groups based on 
coarse overstory groups, and \sQuote{maxGranularity} groups into what is deemed a maximum realistic granluarity 
given the the number of cases of each ecoType. See Details below for the specifics of each grouping. }}

\item{type.of.data}{one of \code{c('transform','labels','colours'), where: \sQuote{transform} is the lookup vector,
\sQuote{labels} are meaningful names for each group, and \sQuote{colours} are suggest plotting colours. }}
}
\description{
This dataset ia a list, cast in matrix format, that can be used to lump ecotypes into coarser groups for analysis and plotting.
  The current implementation includes vectors for: transformation, nameing, and plotting colours
}
\details{
This default transformation 'function' maps several different grouping scenarios. Each is provided with labels for the groups 
  as well as suggest colours, although in the case of identity this is merely the first \code{N} colours. 
\tabular{lccc}{
\bold{Ecotype} \tab \bold{Dominant Species} \tab \bold{Dominant Group} \tab \bold{Maximum Granularity}\cr
BS 01 – Sand heather \tab 1 (barren) \tab 1 (barren) \tab 1 (sparse vegitation)\cr
BS 02 – Lichen felsenmeer \tab 1 \tab 1 \tab 1\cr
BS 03 – Jack pine, blueberry, lichen \tab 2 (pine) \tab 2 (conifer) \tab 2\cr
BS 04 – Jack pine, black spruce, feathermoss \tab 2 \tab 2 \tab 3\cr
BS 05 – Jack pine, birch, feathermoss \tab 2 \tab 2 \tab 4\cr
BS 06 – Jack pine, aspen, alder \tab 2 \tab 2 \tab 5\cr
BS 07 – Black spruce, blueberry, lichen \tab 3 (bs) \tab 2 \tab 6\cr
BS 08 – Black spruce, birch, lichen \tab 3 \tab 2 \tab 7\cr
BS 09 – Black spruce, pine, feathermoss \tab 3 \tab 2 \tab 8\cr
BS 10 – Black spruce, birch, feathermoss \tab 3 \tab 2 \tab 9\cr
BS 11 – White spruce, fir, feathermoss \tab 4 (ws) \tab 2 \tab 10\cr
BS 12 – White spruce, crowberry, feathermoss \tab 4 \tab 2 \tab 11\cr
BS 13 – Birch, black spruce, aspen \tab 5 (birch) \tab 3 (decid) \tab 12\cr
BS 14 – Birch, lingonberry, lab tea \tab 5 \tab 3 \tab 13\cr
BS 15 – Aspen, birch, alder \tab 6 (aspen) \tab 3 \tab 14\cr
BS 16 – Black spruce, poplar, alder swamp \tab 7 (swamp) \tab 4 (wetland) \tab 15 (group swamp in with treed bog)\cr
BS 17 – Black spruce bog \tab 8 (bog) \tab 4 \tab 15\cr
BS 18 – Lab tea shrubby bog \tab 8 \tab 4 \tab 16\cr
BS 19 – Graminoid bog \tab 8 \tab 4 \tab 17 (sparse bog)\cr
BS 20 – Open bog \tab 8 \tab 4 \tab 17\cr
BS 21 – Tamarack fen \tab 9 (fen) \tab 4 \tab 18 (upright fen)\cr
BS 22 – Leatherleaf fen \tab 9 \tab 4 \tab 18\cr
BS 23 – Willow shrubby fen \tab 9 \tab 4 \tab 19\cr
BS 24 – Graminoid fen \tab 9 \tab 4 \tab 20 (sparse fen)\cr
BS 25 – Open fen \tab 9 \tab 4 \tab 20\cr
BS 26 – Rush sandy shore \tab 10 (shore) \tab 4 \tab 21 (shoreline)\cr
BS 27 – Sedge rocky shore \tab 10 \tab 4 \tab 21\cr
}
}
\section{Warning}{

The following expressions will not work as transformation functions. See \code{\link{factor}}, esp. the 'Warning' section.
\describe{
  \item{\code{vData$Ecotype}}{is a factor list, not a numeric list representing the factors}
  \item{\code{ecoGroup[['identity','transform']][vData$Ecotype]}}{will regroup based on the factor level \emph{indices}, NOT the factors}
  \item{\code{as.numeric(vData$Ecotype]}}{returns the factor \emph{indices}, not a numeric representation of the factors}
  \item{\code{(1:27)[vData$Ecotype]}}{returns the factor \emph{index} rather than the ecoGroup} }
}
\examples{
ecoGroup[['domSpecies','transform']][ as.numeric(as.character(vData$Ecotype)) ]
}
\keyword{datasets}

