% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{extractPoints}
\alias{extractPoints}
\title{Extracts the raster data corresponding to each point in the point dataset.}
\usage{
extractPoints(rData, vData, colNames = NULL, na.omit = T)
}
\arguments{
\item{rData}{the raster data brick}

\item{vData}{the point data}

\item{colNames}{(optional) a collection of column names. If unspecified the function will pull them from the layers in the brick.}

\item{na.omit}{remove NA points from the dataset?}
}
\value{
a data frame with the extracted raster data columns appended
}
\description{
Extracts the raster data corresponding to each point in the point dataset.
}

