% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{factorValues}
\alias{factorValues}
\title{Retrieve the values within a factor object}
\usage{
factorValues(x)
}
\description{
This function retrieves the *values* within a factor object. There is a subtle but serious gotcha when using factors
that is outlined in the help for factor() under the Warning section. Notably, then using factors, what is returned
is the *indicies* of the factor, not the values stored in the factor. This is fairly easy to catch when factors are
type character, but can cause serious bugs when they are numeric. The purpose of this function is to provide an easy
way to return the values stored within a factor
}

