% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_util.R
\name{getArgs}
\alias{getArgs}
\title{Extract the arguments used to generate a model object}
\usage{
getArgs(model)
}
\arguments{
\item{model}{is the model for which to extract the arguments}
}
\value{
a list of arguments
}
\description{
This function will recover the arguments used when generating this model object. It does not recover all arguments, but merely those
relating to this model object.
}
\note{
The nearest neighbour models in package:FNN and package:class do not enclose their results in a class; when NPEL.Classification
  builds objects of these types it wraps them in a class so they are recognizable by S3 methods, and attaches the formula and data.
  Hence, if a model was built directly using these packages, the result will not run this function.
}

