% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OOP_util.R
\name{getFitted}
\alias{getFitted}
\title{Extract the fitted data from a model object}
\usage{
getFitted(model)
}
\arguments{
\item{model}{is the model for which to extract the fitted data}
}
\value{
a factor variable containing the fitted data
}
\description{
This function returns the fitted data that the model was built on; that is, if we 'predict' this model using the same
data we used to generate the model, we would get this as a result.
}
\note{
The nearest neighbour models in package:FNN and package:class do not enclose their results in a class; when
  NPEL.Classification builds objects of these types it wraps them in a class so they are recognizable by S3 methods,
  and attaches the formula and data. Hence, if a model was built directly using these packages the result will not
  run this function.
}

