% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_util.R
\name{getFormula}
\alias{getFormula}
\title{Extract the formula used to generate a model object}
\usage{
getFormula(model)
}
\arguments{
\item{model}{is the model for which to extract the formula.}
}
\value{
a "formula" object representing the parameters used to generate the model.
}
\description{
This function will return a formula object from the specified model. The underlying code here uses only the model
object, and casts the result into a formula object even if the model does not natively store the formula in that
format. As a result this function should for all model types this package uses, even if it wasn't built by
generateModels.
}
\note{
The nearest neighbour models in package:FNN and package:class do not enclose their results in a class; when
  NPEL.Classification builds objects of these types it wraps them in a class so they are recognizable by S3 methods,
  and attaches the formula and data. Hence, if a model was built directly using these packages, the result will not
  run this function.
}
