% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_util.R
\name{getVIMP}
\alias{getVIMP}
\title{Extract the VIMP if it is present}
\usage{
getVIMP(model)
}
\arguments{
\item{model}{is the model for which to extract the VIMP}
}
\value{
a data frame with the VIMP data if it is present, otherwise NULL
}
\description{
This function will extract the Variable Importance data if the model contains it, otherwise it attempts to compute it using a
leave-one-out algorithm.
}
\note{
The nearest neighbour models in package:FNN and package:class do not enclose their results in a class; when NPEL.Classification
  builds objects of these types it wraps them in a class so they are recognizable by S3 methods, and attaches the formula and data.
  Hence, if a model was built directly using these packages, the result will not run this function.
}

