% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_util.R
\name{isCat}
\alias{isCat}
\alias{isCont}
\title{Is the object categorical or continuous}
\usage{
isCat(x)

isCont(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
TRUE if the object is categorical/continuous, or is a model built on categorical/continuous data, FALSE otherwise
}
\description{
Checks if an object is categorical (or continuous). Currently accepts either a vector, presumably from a data frame,
or a model object. In the case of the latter, it will check whether the model was built using categorical data.
}
\details{
Note that it cannot be assume that if a model fails \code{isCat} it will pass \code{\link{isCont}}; both functions test for the
conditions which it expects to hold, and there may be cases in which neither set of conditions is fully met.
}
\examples{
isCat(siteData$ecoType)
isCat(siteData$slope)
\dontrun{
isCat(model) }

isCont(siteData$ecoType)
isCont(siteData$slope)
\dontrun{
isCont(model) }
}

