% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{merge.levels}
\alias{merge.levels}
\title{Merge the levels of two factors; the factors of y not present in x will be added to the x levels.}
\usage{
\method{merge}{levels}(x, y)
}
\arguments{
\item{x}{the factor to process}
}
\value{
the factor x with the levels of y merged to the x levels
}
\description{
Merge the levels of two factors; the factors of y not present in x will be added to the x levels.
}

