% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{mergeLevels}
\alias{mergeLevels}
\title{Merge the levels of two factors; the factors of y not present in x will be added to the x levels.}
\usage{
mergeLevels(x, y)
}
\arguments{
\item{x}{the factor to which a the levels of...}

\item{y}{the factor for which to extract the added levels}
}
\value{
the factor x with the levels of y merged to the x levels
}
\description{
Merge the levels of two factors; the factors of y not present in x will be added to the x levels.
}
