% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwTile.R
\name{mergeTile}
\alias{mergeTile}
\title{Merge a list of raster objects into a single brick}
\usage{
mergeTile(dataList, path = "", fileName = "")
}
\arguments{
\item{dataList}{a list of raster* objects}

\item{path}{(optional) path to the output file. If unspecified, getwd() will be used.}

\item{fileName}{(optional) output filename to associate with the new brick. If unspecified, the filename of the first element will be used.}
}
\value{
a brick with the specified data
}
\description{
A utility function to merge a list of raster* objects and output them as a brick. Note that all
the objects will need to obey the limitations of brick or this function will fail.
}

