% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{minExtent}
\alias{minExtent}
\title{Find the minimum extent of a collection of Raster* object}
\usage{
minExtent(...)
}
\arguments{
\item{...}{the Raster* objects to compare}
}
\value{
the minimum extent of all Raster* objects
}
\description{
Find the minimum extent of a collection of Raster* object
}

