% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Models.R
\name{modelAccs}
\alias{modelAccs}
\title{Generate accuracy statistics and variable importance for a list of models}
\usage{
modelAccs(models, classNames = NULL, calc = F, echo = T)
}
\arguments{
\item{models}{is either a list of model objects on which to find error statistics, or a single model to evaluate}

\item{classNames}{(optional) class names to attach to the tables}

\item{calc}{(optional) passed on to \code{\link{npelVIMP}}, defaults to false}

\item{echo}{(optional) should the function print out it's progress, defaults to false}
}
\value{
A named list of accuracy and VIMP data: \code{confMatrix} = confusion matrix, \code{userAcc} = user accuracy, \code{prodAcc} =
  producer accuracy, \code{overallAcc} = overall accuracy, \code{kappa} = kappa, \code{VIMP} = VIMP matrices, \code{VIMPoverall} =
  overall VIMP
}
\description{
Given a block (list) of models, this function runs \code{\link{classAcc}} and the \code{\link{npelVIMP}} metric on each. See details for a
description of how the data is encapsulated for return.
}
\details{
The function returns a named list of the various accuracy and variable importance metrics:
\itemize{
  \item \code{Accuracies} contains the complete names list returned by \code{\link{classAcc}}; see the help on that method for more
    information on that data structure.
  \item \code{VIMP} a named list containing the variable importance matrices (VIMP) for each included model. Each entry in the list is a
    data frame in which columns are different classes and rows are different input variables; specific entries represent the variable
    importance of a given input variable on a particular class.
  \item \code{VIMPoverall} a data.frame showing the overall variable importance (VIMP) for a given model.

Columns are different models and entries in the table are the standardized VIMP for an input variable on that model. Values have been
  standardized by column so they are (somewhat) comparable as the algorithm by which VIMP is computed varies by model--hence the largest
  value will be 1 in every column. This makes the columns approximately comparable, however, values should be taken with a grain of salt,
  and perhaps rank order is the most robust comparison.
}
}
\note{
See the details section of \code{\link{npelVIMP}} for a discussion of the limitations of our VIMP metric.
}
\examples{
data ('siteData')
modelRun <- generateModels (data = siteData,
                            modelTypes = c ('rF','rFSRC','fnn.FNN','fnn.class','kknn','gbm'),
                            x = c('brtns','grnns','wetns','dem','slp','asp','hsd'),
                            y = 'ecoType',
                            grouping = ecoGroup[['domSpecies','transform']])
mE <- modelAccs (modelRun,ecoGroup[['identity','labels']])
}
\seealso{
\code{\link{classAcc}} for more on class level accuracies, and \code{\link{npelVIMP}} for more on how that metric is computed and it's limitations.
}

