% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{nnVIMP}
\alias{nnVIMP}
\title{Generate an estimate for variable importance for nearest neighbour models}
\usage{
nnVIMP(model)
}
\arguments{
\item{model}{is the nearest neighbour model to test}
}
\description{
In an attempt to establish some indication of variable importance for nearest neighbour models, this algorithm computes the overall accuracies of the specified
  model successively leaving out one predictor variable. The VIMP score is then calculated by linearly scaling the overall accuracies such that the version with
  the largest drop in overall accuracy scores 1.0, and the model with the smallest drop in overall accuracy scores 0.0
}
\details{
The function returns a named vector of the VIMP scores for each variable
}

