% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{npelVIF}
\alias{npelVIF}
\title{Compute the variable inflation factor of the terms in a formula}
\usage{
npelVIF(fx, data)
}
\arguments{
\item{fx}{a formula for which to test the VIF}

\item{data}{a data frame with columns corresponding to the formula}
}
\value{
a list with the following elements: list (vif, rSquared, est, stdErr, t.value, p.value). See Details
}
\description{
Given an input dataset, this function generates the VIF for each predictor variable. It gives considerably more output than the typical
VIF function---see details.
}
\details{
The function returns a named list of several details from the computation of the VIF scores :
\itemize{
  \item \code{vif} the resulting VIF scores for each term in the model.
  \item \code{rSquared} the resulting R-square with that term \emph{removed} from the model.
  \item \code{est} a matrix of parameter estimates from each linear model. Rows are for each parameter that was estimated, and columns
  are the parameter estimates. Note that the parameter that is being estimated is represented by NA in the estimates since it was not
  included in the model.
  \item \code{stdErr} a matrix of the standard error for each parameter estimate in each model; the format is the same as for \code{est}.
  \item \code{t.value} the t.value for each parameter estimate in each model; the format is the same as for \code{est}.
  \item \code{p.value} the p.value for each parameter estimate in each model; the format is the same as for \code{est}.
}
}
\examples{
data (siteData)
npelVIF(formula('Ecotype ~ brtns+grnns+wetns+dem+slp+asp+hsd'),siteData)
}
