% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotTile.base}
\alias{plotTile.base}
\title{Creates a base ggplot object of the raster object}
\usage{
plotTile.base(data, layers = NULL, title = "", maxpixels = 5e+05,
  reduction = 1, ...)
}
\arguments{
\item{data}{a raster* object to plot}

\item{layers}{(optional) the layers to include as facets}

\item{title}{(optional) a title for the plot}

\item{maxpixels}{(optional) the maximum number of pixels to output}

\item{reduction}{(optional) is a reducing scale factor applied to both x and y axis; will use whichever is less: ncell/(reduction^2) or maxpixels}

\item{...}{other parameters for ggplot}
}
\value{
a ggplot object which can have geoms, scales, etc. added to it (see PlotTile).
}
\description{
Creates a base ggplot object of the raster object
}

