% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotTile}
\alias{plotTile}
\title{Given a ggplot base object create a finished plot}
\usage{
plotTile(gp, layers, discrete, colours, labels = NULL, ...)
}
\arguments{
\item{gp}{the base ggplot to which to add a colour scale and aesthetic (see plotTile.base)}

\item{layers}{a character vector of the levels which to plot}

\item{discrete}{force the plot to attempt a discrete fill axis}

\item{colours}{specify the colours use; a list if discrete is TRUE, or, 2 or 3 colours for the gradient if discrete is FALSE}

\item{labels}{if desired, specify the names of the labels for each class}

\item{...}{other parameters to pass ???}
}
\value{
the ggplot object with additional aes and scale
}
\description{
Given a ggplot base object create a finished plot
}

