% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotTiles}
\alias{plotTiles}
\title{Outputs plots of multiple data files into a single folder}
\usage{
plotTiles(path, base, extension, models, type = "pdf", ...)
}
\arguments{
\item{path}{the path of the folder}

\item{base}{the base part of the filename}

\item{extension}{the filename 'extension'; may contain more than the strict extension}

\item{models}{a list of model names}

\item{type}{(optional) type out output to use}

\item{...}{other parameters to pass to the device function}
}
\description{
Outputs plots of multiple data files into a single folder
}
