% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwTile.R
\name{readTile}
\alias{readTile}
\title{Read a collection of layers}
\usage{
readTile(rasterPath, rasterName, layers, labels = NULL, NAval = NULL)
}
\arguments{
\item{rasterPath}{path to the raster file}

\item{rasterName}{name of the file}

\item{layers}{a character list of layers to read. The filename used will be 'rasterName_layer.tif'}

\item{labels}{(optional) a list of names for the columns of the final raster. Will attempt to use the layer names if it appears there is only one layer per file.}

\item{NAval}{the value to be interpreted as NA; passed to raster::NAvalue}
}
\value{
a raster.stack of the objects in the folder
}
\description{
The function reads a collection of geotiffs and creates a single raster.stack object
}

