% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Docs.R
\docType{data}
\name{siteData}
\alias{#}
\alias{???}
\alias{siteData}
\title{An example of site data; this is the portion of the data collected in-house.}
\format{A data frame with columns as variables and rows as sites.}
\source{
\itemize{
  \item \bold{Landsat data}: available from the U.S. Geological Survey
  \item \bold{DEM data}: available from the U.S. Geological Survey National Elevation Dataset
}
}
\usage{
siteData$'variables'
}
\arguments{
\item{variables}{is the variable to extract from the dataset}
}
\description{
This data frame contains the site data collected in-house, with the corresponding remote sensed data, both Landsat and DEM and their
respective derived categories. See the details section for more information on each included variable.
}
\details{
In order to build classification models the NPEL collected field data. This vector (point) data was overlaid on the raster data, and the
value at each point (site) was extracted. This yielded a dataset of sites with both field characteristics and remote sensed data. This dataset
is the basis for the model building in this package. What is included here is a subset both of variables and sites that can be used for
learning and testing this package. The sites are only those collected in-house at the NPEL; because our project is done in collaboration with
several other agencies we combined their data with ours to produce a larger dataset that spanned more geographic area and more ecosite types.
Many of the variable s collected are not relevant to building models (e.g. plot number, date, etc.) and have been pruned from this example. Also
several of the remote sensed layers did not prove to be sufficiently linearly independent enough to warrant inclusion in the model; several
redundant variables are included here for the purposes of demonstrating dimensional analysis but some have been trimmed in order to keep the
example manageable.

\itemize{
  \item \code{block} the block it was sampled under, a component of the sampling regime: c('LA','B1','B2',...,'B6') -- large area, blocks 1..6
  \item \code{landClass} the pre-sampling classification of this site, i.e. the stratum it was sampled under
  \item \code{GPSzone} UTM zone of this site; 13 for all our sites
  \item \code{easting} UTM easting of this site
  \item \code{northing} UTM northing of this site
  \item \code{year} the year the field site was visited
  \item \code{slope} field recorded slope of the site, useful to compare with DEM data
  \item \code{aspect} field recorded aspect of the site, useful to compare with DEM data
  \item \code{lfhD} depth of LFH (litter, fibric, humic) soil layer
  \item \code{waterD} depth to water
  \item \code{bedrockD} depth to bedrock
  \item \code{parentMaterial} the type of parent material (soil) found at the site
  \item \code{effTexture} the effective texture of the soil
  \item \code{ecoType} the ecotype the site was classed as
  \item \code{base.1:7} Landsat base layers
  \item \code{evi} enhanced vegetation index -- derived from Landsat data
  \item \code{ndvi} normalized difference vegetation index -- derived from Landsat data
  \item \code{ndwi} normalized difference water index -- derived from Landsat data
  \item \code{mavi} modified soil-adjusted vegetation index -- derived from Landsat data
  \item \code{grnns, wetness, brtns} the so-called \sQuote{Tasseled-Cap} variables; they reflect greenness, wetness, and brightness respectively
  \item \code{dem} the basic elevations from the digital elevation model
  \item \code{slp} slope, as derived from the DEM
  \item \code{asp} aspect, as derived from the DEM; note that different algorithms use different values for NA, a reliable filter is that aspect is NA when slope == 0
  \item \code{srr} surface relief ratio, a measure of surface roughness -- derived from the DEM
  \item \code{cti} compound topographic index, a measure of surface roughness -- derived from the DEM
  \item \code{ri} roughness index, a measure of surface roughness -- derived from the DEM
  \item \code{hsd} hillshade, a measure of how much light the surface receives -- derived from the DEM
}
}
\examples{
library (NPEL.Classification)
data(siteData)
names(siteData)

siteData$ecoType
siteData[,'ecoType']
siteData[,c('easting','northing')]

modelRun <- generateModels (data=siteData,
                            modelTypes=c('rF','rFSRC','fnn.FNN','fnn.class','kknn','gbm'),
                            fx=formula('ecoType ~ brtns+grnns+wetns+dem+slp+asp+hsd'),
                            grouping=ecoGroup[['domSpecies','transform']])

\dontrun{
vData <- maptools::readShapePoints ('Input/Plots/Plots')         # Read in from a ESRI shapefile
vData$FEC <- as.factor(substr(as.character(vData$FEC),3,6))      # Extract the FEC number
names (vData) <- c('objID','plotNum','block','type','landClass','ecoRegion','GPSzone','etc..') }
}
\seealso{
See \code{\link{extractPoints}} for a function to extract raster data for these sites, and \code{\link{generateModels}} for model building.
}
\keyword{datasets}

