% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{sort.levels}
\alias{sort.levels}
\title{Sort the levels in a factor}
\usage{
\method{sort}{levels}(x)
}
\arguments{
\item{x}{the factor to process}
}
\value{
the same data with levels sorted
}
\description{
Sort the levels in a factor
}

