% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{sortLevels}
\alias{sortLevels}
\title{Sort the levels in a factor}
\usage{
sortLevels(x)
}
\arguments{
\item{x}{the factor to process}
}
\value{
the same factor variable with levels sorted
}
\description{
Sort the levels in a factor
}
