% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{trim.levels}
\alias{trim.levels}
\title{Trim the levels so the only levels present are ones that actually have data}
\usage{
trim.levels(x)
}
\arguments{
\item{x}{the factor to process}
}
\value{
the same data with the levels trimmed to reflect the factors that are present
}
\description{
Trim the levels so the only levels present are ones that actually have data
}

