% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{validate}
\alias{validate}
\title{Check the accuracy of a model using independent validation data}
\usage{
validate(model, valid, ...)
}
\arguments{
\item{model}{the model to test.}

\item{valid}{the validation dataset; must contain all the parameters used in the model.}

\item{...}{other variables to pass to \code{\link{classAcc}}.}
}
\value{
Returns the result from \code{\link{classAcc}}
}
\description{
This function is a wrapper for classAcc that, given an independent validation dataset, uses the model to come up with new predicted
values. It then calls \code{\link{classAcc}} with these two validation datasets (the one generated by the model and the true values) to
report class accuracies.
}
\examples{
data ('siteData')
# With categorical data
gen <- sample(1:nrow(siteData),floor(nrow(siteData)*0.5))
modelRun <- generateModels(data = siteData[sort(gen),],
                           modelTypes = suppModels,
                           x = c('brtns','grnns','wetns','dem','slp','asp','hsd'),
                           y = 'ecoType',
                           grouping = ecoGroup[['identity','transform']],
                           echo = FALSE)
valid <- siteData[-gen,]
valid$ecoType <- as.factor(ecoGroup[['identity','transform']][valid$ecoType])
validate(modelRun[[2]],valid)
validModels(modelRun,valid)

# With continuous data
gen <- sample(1:nrow(siteData),floor(nrow(siteData)*0.5))
modelRun <- generateModels(data = siteData[sort(gen),],
                           modelTypes = contModels,
                           x = c('brtns','grnns','wetns','dem','slp','asp','hsd'),
                           y = 'easting',
                           echo = FALSE)
valid <- siteData[-gen,]
validate(modelRun[[2]],valid)
modelsValid(modelRun,valid)
}
\seealso{
\code{\link{classAcc}}
}

