% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\docType{data}
\name{water}
\alias{water}
\title{An example of a water mask}
\format{A Raster.Layer object with an example water mask: values of 1 indicate water, values of 128 indicate non-water.}
\usage{
water
}
\description{
Since it is unlikely to have field sites situated on water, most classification models are not well equipped to handle water areas.
Typically this is handled by simply ignoring them when building the model. However, depending on the post-rendering analysis, it can be
useful to mask them out. This \code{Raster.Layer} is a sample water mask for the sample dataset provided \code{\link{egTile}}. See the
examples section for how to use this mask.
}
\examples{
data(egTile)
data(water)
plot (egTile)
plot (raster::mask(egTile,water,maskvalue=1,updatevalue=NA))
}
\seealso{
See \code{\link{writeTile}} and \code{\link{writeTiles}} for examples of rendering.
}
\keyword{datasets}

