% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{wrapAccs.Cat}
\alias{wrapAccs.Cat}
\alias{wrapAccs.Cont}
\title{Thin wrappers to package up categorical/continous accuracies/validation data for a group of models.}
\usage{
wrapAccs.Cat(models, valid, func, VIMP, calc, classNames, echo)

wrapAccs.Cont(models, valid, func, VIMP, calc, echo)
}
\arguments{
\item{models}{the model group to process.}

\item{valid}{validation data if relevant, i.e. if func = 'validate'.}

\item{func}{the function to use to evaluate accuracy: classAcc or validate.}

\item{VIMP}{should we add the VIMP data.}

\item{calc}{should we compute the VIMP metric from scratch, only referenced if VIMP = TRUE.}

\item{classNames}{the classnames.}

\item{echo}{should the function print out processing messages as it runs.}
}
\value{
a the class that \code{\link{modelAccs}} or \code{\link{modelValid}} needs.
}
\description{
These functions are intended for internal use---as such no defaults are provided and there is little to no internal parameter
checking. Using this function it is possible, however, to do things like recover VIMP data using a validating dataset, which is not
possible using the default package functions.
}

