% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwTile.R
\name{writeTile}
\alias{writeTile}
\title{Computes an output landscape}
\usage{
writeTile(outFilename, inImage, model, layers = c("class"), threshold = 0,
  n.all = 0, labels.all = NULL, ...)
}
\arguments{
\item{outFilename}{the output filename}

\item{inImage}{a raster* of the input data}

\item{model}{the model to use for predicting output values. The model class
must have a predict.model function.}

\item{layers}{one of: "class" the chosen class; "prob" the probability of
this class; "threshold" same as class where probability is greater than
or equal to threshold (below), 0 otherwise; "all" the probabilies for all
classes (one layer per possibility). Note: this may be a large dataset!}

\item{threshold}{if the "threshold" layer is selected, this needs to be set to the threshold probability (0:1).}

\item{n.all}{if the 'all' group is selected in layers, specify how many layers this is going to be.}

\item{labels.all}{if the 'all' group is selected in layers, specify labels for these layers.}

\item{...}{other variables to pass to the model predict function.}
}
\value{
the new raster* object with the output landscape
}
\description{
This function computes and outputs a generated landscape, and optionally the probabilities
  of the winning class and/or every possible class.
}
\details{
Note: outputing all the probabilities can generate quite a large datafile, that is, in the order of a GB or more.
}

