% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwTile.R
\name{writeTiles}
\alias{writeTiles}
\title{Write multiple output tiles}
\usage{
writeTiles(path, base, extension, modelRun, models, layers = c("class"))
}
\arguments{
\item{path}{the path of the folder}

\item{base}{the base part of the filename}

\item{models}{a list of model names}

\item{layers}{one or more of 'class':the output classes, 'prob':the probability of that class, 'all': the probability for every possible class}

\item{ext}{the latter part of the filename}
}
\value{
the time it took to execute
}
\description{
A utility function to output multiple datafiles in a single folder.
}
\details{
The output filename is <path><base><model type><ext>
}

