% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{classAcc}
\alias{classAcc}
\title{Generates basic error statistics for a classification model}
\usage{
classAcc(pred, valid, digits = 3, classNames = NULL)
}
\arguments{
\item{digits}{(optional) the number of digits to output for the error}

\item{x}{the prediction classes}

\item{y}{the validation classes}
}
\value{
the error data as a five element list: confusion matrix, user accuracy, producer accuracy, overall accuracy, kappa
}
\description{
Generates basic error statistics for a classification model
}
\details{
Create some notes here on how each accuracy rate is computed and it's significance???
}

