% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(x)
}
\arguments{
\item{x}{the input angle (in degrees)}
}
\value{
The angle in radians
}
\description{
Convert degrees to radians
}

