% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Docs.R
\docType{data}
\name{ecoGroup}
\alias{ecoGroup}
\title{An example of a transformation 'function'}
\format{A list cast into matrix format with four rows (grouping scenarios) and three columns (possible data types for
  that scenario). The four different grouping scenarios are:
\itemize{
  \item \bold{Identity} (\sQuote{\code{identity}}) maps the class onto itself, that is, there is no lumping, but
    labels and colours are provided.
  \item \bold{Dominant Species} (\sQuote{\code{domSpecies}}) groups classes by the dominant (tree) species.
  \item \bold{Dominant Group} (\sQuote{\code{domGroup}}) groups classes by the dominant growth-form.
  \item \bold{Maximum Granularity} (\sQuote{\code{maxGranularity}}) groups the classes in what is deemed reasonable
    for our data, that is, we attempt to maximize the number of groupings while maintaining a reasonable number of
    samples in each group for statistical significance and reliability.}

The different types of data that can be extracted are:
\itemize{
  \item\bold{Transformation Vector} (\sQuote{\code{transform}}) a transformation vector converting input factors
    (ecoSites) to output factors (groupings). Functionally it is a lookup table that maps input classes to output
    classes. See \sQuote{examples} for how to use it.
  \item\bold{Labels} (\sQuote{\code{labels}}) meaningful names for the output groups.
  \item\bold{Plotting Colours} (\sQuote{\code{colours}}) suggested plotting colours for the output groups. These
    are easy to pass to, for example, plotting functions.}}
\usage{
ecoGroup[[which.grouping,type.of.data]]
}
\arguments{
\item{which.grouping}{one of \code{c('identity','domSpecies','domGroup','maxGranularity')}, see Format.}

\item{type.of.data}{one of \code{c('transform','labels','colours')}, see Format.}
}
\description{
This dataset is a list, cast in matrix format, that can be used to lump ecotypes into coarser groups for analysis and
plotting. The current implementation includes vectors for: transformation, naming/labelling, and suggested plotting
colours
}
\details{
This built-in transformation 'function' encodes several different grouping scenarios. Each is provided with labels
for the groups as well as suggested colours. Following is a description of the four build-in encoding for the FEC
classification system \cite{McLaughlan, M. S., Robert A. Wright, and R. D. Jiricka. Field Guide to the Ecosites of
Saskatchewan's Provincial Forests. Prince Albert, Sask: Ministry of Environment, 2010.}

\tabular{lccc}{
\bold{Ecotype} \tab\bold{Dominant Species Class} \tab\bold{Dominant Group Class} \tab\bold{Maximum Granularity Class}\cr
  BS 01 -- Sand heather \tab 1 (barren) \tab 1 (barren) \tab 1 (sparse vegetation)\cr
  BS 02 -- Lichen felsenmeer \tab 1 \tab 1 \tab 1\cr
  BS 03 -- Jack pine, blueberry, lichen \tab 2 (pine) \tab 2 (conifer) \tab 2\cr
  BS 04 -- Jack pine, black spruce, feathermoss \tab 2 \tab 2 \tab 3\cr
  BS 05 -- Jack pine, birch, feathermoss \tab 2 \tab 2 \tab 4\cr
  BS 06 -- Jack pine, aspen, alder \tab 2 \tab 2 \tab 5\cr
  BS 07 -- Black spruce, blueberry, lichen \tab 3 (bs) \tab 2 \tab 6\cr
  BS 08 -- Black spruce, birch, lichen \tab 3 \tab 2 \tab 7\cr
  BS 09 -- Black spruce, pine, feathermoss \tab 3 \tab 2 \tab 8\cr
  BS 10 -- Black spruce, birch, feathermoss \tab 3 \tab 2 \tab 9\cr
  BS 11 -- White spruce, fir, feathermoss \tab 4 (ws) \tab 2 \tab 10\cr
  BS 12 -- White spruce, crowberry, feathermoss \tab 4 \tab 2 \tab 11\cr
  BS 13 -- Birch, black spruce, aspen \tab 5 (birch) \tab 3 (decid) \tab 12\cr
  BS 14 -- Birch, lingonberry, lab tea \tab 5 \tab 3 \tab 13\cr
  BS 15 -- Aspen, birch, alder \tab 6 (aspen) \tab 3 \tab 14\cr
  BS 16 -- Black spruce, poplar, alder swamp \tab 7 (swamp) \tab 4 (wetland) \tab 15 (group swamp in with treed bog)\cr
  BS 17 -- Black spruce bog \tab 8 (bog) \tab 4 \tab 15\cr
  BS 18 -- Lab tea shrubby bog \tab 8 \tab 4 \tab 16\cr
  BS 19 -- Graminoid bog \tab 8 \tab 4 \tab 17 (sparse bog)\cr
  BS 20 -- Open bog \tab 8 \tab 4 \tab 17\cr
  BS 21 -- Tamarack fen \tab 9 (fen) \tab 4 \tab 18 (upright fen)\cr
  BS 22 -- Leatherleaf fen \tab 9 \tab 4 \tab 18\cr
  BS 23 -- Willow shrubby fen \tab 9 \tab 4 \tab 19\cr
  BS 24 -- Graminoid fen \tab 9 \tab 4 \tab 20 (sparse fen)\cr
  BS 25 -- Open fen \tab 9 \tab 4 \tab 20\cr
  BS 26 -- Rush sandy shore \tab 10 (shore) \tab 4 \tab 21 (shoreline)\cr
  BS 27 -- Sedge rocky shore \tab 10 \tab 4 \tab 21\cr
}
}
\section{Warning}{

The following expressions will \emph{not} work as transformation functions. See \code{\link{factor}}, especially the \code{Warning}
section for more information on this gotcha.

\code{siteData$Ecotype} returns a factor list, not a numeric list representing the factors.

\code{ecoGroup[['identity','transform']][siteData$Ecotype]} will regroup based on the factor level \emph{indices}, NOT the factors.

\code{as.numeric(siteData$Ecotype]} returns the factor \emph{indices}, not a numeric representation of the factors.

\code{(1:27)[siteData$Ecotype]} returns the factor \emph{index} rather than the ecoGroup.
}
\examples{
data (siteData)
ecoGroup[['domSpecies','transform']][ factorValues(siteData$Ecotype) ]

# Code for generating this object--so you can make your own!
# Create a shell to fill with data
ecoGroup <- list(); length(ecoGroup) <- 12;
dim(ecoGroup) <- c(4,3)                             # Three rows (different classifiers), and three columns (type of data we could extract)
rownames(ecoGroup)=c('identity','domSpecies','maxGranularity','domGroup') # Rows are different grouping scenarios
colnames(ecoGroup)=c('transform','labels','colours')                      # Cols are different things we could look up

# Fill with data
ecoGroup['identity',] <- list(c(1:27), paste0('BS',1:27), 1:27)
ecoGroup['domSpecies',] <- list( c(1,1,2,2,2,2,3,3,3,3,4,4,5,5,6,7,8,8,8,8,9,9,9,9,9,10,10),
                                 c('1'="Barren",'2'="Pine",'3'="Black Sp",'4'="White Sp",'5'="Birch",'6'="Aspen",'7'="Swamp",'8'="Bog",'9'="Fen",'10'="Shore"),
                                 c('1'="gray",'2'="tan1",'3'="green4",'4'="seagreen",'5'="green3",'6'="green2",'7'="steelblue",'8'="steelblue2",'9'="steelblue4",'10'="azure") )
ecoGroup['maxGranularity',] <- list( c(1,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,15,16,17,17,18,18,19,20,20,21,21), paste0('BS',1:21), 1:21)
ecoGroup['domGroup',] <- list( c(1,1,2,2,2,2,2,2,2,2,2,2,3,3,3,4,4,4,4,4,4,4,4,4,4,4,4),
                               c("Barren","Conifer","Decid","Wetland"),
                               c('1'="gray",'2'="green4",'3'="green2",'4'="steelblue") )
}
\keyword{datasets}

