% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwTile.R
\name{extractPoints}
\alias{extractPoints}
\title{Extracts the raster data corresponding to each point in the vector dataset.}
\usage{
extractPoints(rData, vData, locs, na.omit = TRUE)
}
\arguments{
\item{rData}{the raster data as a raster.* object; can be read in using \code{\link{readTile}}}

\item{vData}{the point data as a data frame or a SpatialPointsDataFrame, i.e. read in with \code{\link[maptools]{readShapePoints}}}

\item{locs}{the columns which contain the locations\ldots c(easting, northing)}

\item{na.omit}{(optional) remove NA points from the dataset? defaults to TRUE}
}
\value{
a data frame with the combined data: the original vector data with the extracted raster data columns appended
}
\description{
This function looks up each point in a raster (point) dataset and returns the corresponding data from a raster.* object using
sp::SpatialPoints. It is used to build combined remote-sensed and field data that can be input into a classification model.
}
\examples{
egTile <- readTile(file.path(system.file("extdata", "egTile", package = "NPEL.Classification"),''),
                   layers=c('base','grnns','wetns','brtns','dem','slp','asp','hsd'))
vData <- maptools::readShapePoints(system.file("extdata/Plots", "Plots.shp", package = "NPEL.Classification"))
siteData <- extractPoints(egTile,vData,c('EASTING','NORTHING'))
detach ('package:maptools',unload=TRUE)
}
\seealso{
\code{\link{readTile}} for reading in a collection of geoTiff's, \code{\link{egTile}} for more information on the raster format
  and some examples on generating derived raster layers, and \code{\link{siteData}} for an example of the output of this function. See
  also \code{\link[maptools]{readShapePoints}} for more on reading-in vector data from a shapefile.
}

