% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{generateModels}
\alias{generateModels}
\title{A utility function to generate various classification models from a single input dataset}
\usage{
generateModels(data, modelTypes, x = NULL, y = NULL, grouping = NULL,
  echo = F, rf.args = NULL, nn.args = NULL, gbm.args = NULL,
  svm.args = NULL)
}
\arguments{
\item{x}{(optional): vector names of 'predictor' variables to use. Defaults to all columns less the y varible}

\item{y}{(optional): the name of the column of the 'response' variable. Defaults to first column.}

\item{grouping}{(optional): a transformation vector for input classes. Defaults to identity.}

\item{echo}{(optional): should the function report it's progress?}

\item{rf.args}{(optional): arguments to pass to random forest type models}

\item{nn.args}{(optional): arguments to pass to nearest neighbour type models}

\item{gbm.args}{(optional): arguments to pass to gbm}

\item{svm.args}{(optional): arguments to pass to svm}

\item{data:}{the input dataframe}

\item{modelTypes:}{a vector of model types to generate. One or more of c('rF','rFSRC','fnn.FNN','fnn.class','kknn','gbm','svm')}
}
\value{
A named list of models
}
\description{
A utility function to generate various classification models from a single input dataset
}

