% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_util.R
\name{getClasses}
\alias{getClasses}
\title{Recover the classes that a model uses}
\usage{
getClasses(model)
}
\arguments{
\item{model}{is the model from which to recover the classes}
}
\value{
a character vector of classes
}
\description{
This function recovers the classes present in this model That is, it returns the levels used in the original
predicted variable. Of course, this is only relevent if the model was built using categorical data; otherwise
an error will be returned.
}
\note{
The nearest neighbour models in package:FNN and package:class do not enclose their results in a class; when
  NPEL.Classification builds objects of these types it wraps them in a class so they are recognizable by S3 methods,
  and attaches the formula and data. Hence, if a model was built directly using these packages the result will not
  run this function.
}

