% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OOP_util.R
\name{isCat}
\alias{isCat}
\title{Is the object categorical (as opposed to continuous)}
\usage{
isCat(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
TRUE if the object is categorical, or is a model built on categorical data, FALSE otherwise
}
\description{
Checks if an object is categorical (or continuous). Currnetly accepts either a vector, presumably from a data frame,
or a model object. In the case of the latter, it will check whether the model was built using categorical data.
}
\examples{
isCat(siteData$ecoType)
isCat(siteData$slope)
\dontrun{
isCat(model)
}
}
\seealso{
\code{\link{isCont}} for the complementary test.
}

