% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOP_util.R
\name{isCont}
\alias{isCont}
\title{Is the object continuous (as opposed to categorical)}
\usage{
isCont(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
TRUE if the object is continuous, or is a model built on continuous data, FALSE otherwise
}
\description{
Checks if an object is continuous (or categorical). Currnetly accepts either a vector, presumably from a data frame,
or a model object. In the case of the latter, it will check whether the model was built using continuous data.
}
\examples{
isCont(siteData$ecoType)
isCont(siteData$slope)
\dontrun{
isCont(model)
}
}
\seealso{
\code{\link{isCat}} for the complementary test.
}

