% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{modelAccs}
\alias{modelAccs}
\title{Generate accuracy statistics and variable importance for a list of models; a wrapper for \code{classAcc}.}
\usage{
modelAccs(models, classNames = NULL, calc = F, echo = T)
}
\arguments{
\item{models}{is either a list of model objects on which to find error statistics, or a single model to evaluate.}

\item{classNames}{(optional) class names to attach to the tables; only used if the data is categorical.}

\item{calc}{(optional) passed on to \code{\link{npelVIMP}}, defaults to false.}

\item{echo}{(optional) should the function print out it's progress, defaults to false.}
}
\value{
This function returns different values depending on whether the model is categorical or continuous. For categorical data it
  returns a named list of accuracy and VIMP statistics:
  \itemize{
    \item \code{confMatrix} = confusion matrix
    \item \code{userAcc} = user accuracy
    \item \code{prodAcc} = producer accuracy
    \item \code{overallAcc} = overall accuracy
    \item \code{kappa} = kappa
    \item \code{VIMP} = VIMP matrices
    \item \code{VIMPoverall} = overall VIMP
  }
For continuous data:
  \itemize{
    \item \code{overallAcc} = overall r-squared
    \item \code{mse} = mean squared error
  }
}
\description{
Given a list of models, this function runs \code{\link{classAcc}} and \code{\link{npelVIMP}} on each. See details for a
description of how the data is encapsulated for return.
}
\details{
The function returns a named list of the various accuracy and variable importance metrics:
\itemize{
  \item \code{Accuracies} contains the complete names list returned by \code{\link{classAcc}}; see the help on that method for more
    information on that data structure.
  \item \code{VIMP} a named list containing the variable importance matrices (VIMP) for each included model. Each entry in the list is a
    data frame in which columns are different classes and rows are different input variables; specific entries represent the variable
    importance of a given input variable on a particular class.
  \item \code{VIMPoverall} a data.frame showing the overall variable importance (VIMP) for a given model.

Columns are different models and entries in the table are the standardized VIMP for an input variable on that model. Values have been
  standardized by column so they are (somewhat) comparable as the algorithm by which VIMP is computed varies by model--hence the largest
  value will be 1 in every column. This makes the columns approximately comparable, however, values should be taken with a grain of salt,
  and perhaps rank order is the most robust comparison.
}
}
\note{
See the details section of \code{\link{npelVIMP}} for a discussion of the limitations of our VIMP metric.
}
\examples{
# Categorical Data
data ('siteData')
modelRun <- generateModels (data = siteData,
                            modelTypes = suppModels,
                            x = c('brtns','grnns','wetns','dem','slp','asp','hsd'),
                            y = 'ecoType',
                            grouping = ecoGroup[['domSpecies','transform']],
                            echo = FALSE)
mE <- modelAccs (modelRun, ecoGroup[['domSpecies','labels']])
str(mE,1)

# Continuous Data
modelRun <- generateModels (data = siteData,
                            modelTypes = contModels,
                            x = c('brtns','grnns','wetns','dem','slp','asp','hsd'),
                            y = 'easting',
                            echo = FALSE)
mE <- modelAccs (modelRun)
str(mE,1)
}
\seealso{
\code{\link{classAcc}} for more on continuous and categorical accuracies, and \code{\link{npelVIMP}} for more on how that metric
  is computed and it's limitations. Also see \code{\link{modelsValid}} for a function to compute accuracies using an independent dataset.
}

