% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{NPELvif}
\alias{NPELvif}
\title{Compute the VIF of the terms in a formula}
\usage{
NPELvif(formula, data)
}
\arguments{
\item{formula}{a formula}

\item{data}{the input data to use in the regressions}
}
\value{
a list with the following elements: list (vif, rSquared, est, stdErr, t.value, p.value). See Details
}
\description{
Given an input dataset, this function generates VIF output for each term on the left hand side of the specified
  formula. It gives considerable output than the typcial VIF function--see Details.
}
\details{
The function returns a list with more terms than the typical VIF function:
\describe{
  \item{\sQuote{vif}}{the resulting VIF scores for each term in the model}
  \item{\sQuote{rSquared}}{the resulting R-square with that term \emph{removed} from the model}
  \item{\sQuote{est}}{a matrix of parameter estimates from each linear model. Rows are for each parameter
     that was estimated, and columns are the parameter estimates. Not that the parameter that is being
     estimated is represented by a 0 in the estimates since it was not included in the model.}
  \item{\sQuote{stdErr}}{a matrix of the standard error of each parameter estimate. Format is the same as for est.}
  \item{\sQuote{t.value}}{the t.value for each parameter estimate. Format is the same as for est.}
  \item{\sQuote{p.value}}{the p.value for each parameter estimate. Format is the same as for est.}
}
}

