% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Models.R
\name{wrapAccs.Cat}
\alias{wrapAccs.Cat}
\alias{wrapAccs.Cont}
\title{Thin wrappers to package up categorical/continuous accuracies/validation data for a group of models.}
\usage{
wrapAccs.Cat(models, valid, func, VIMP, echo = FALSE, ...)

wrapAccs.Cont(models, valid, func, VIMP, echo = FALSE, ...)
}
\arguments{
\item{models}{the model group to process.}

\item{valid}{validation data if relevant, i.e. if func = 'validate'.}

\item{func}{the function to use to evaluate accuracy: \code{classAcc} or \code{validate}.}

\item{VIMP}{should we append the VIMP data.}

\item{echo}{should the function print out processing messages as it runs.}

\item{...}{(optional) other parameters to pass to \code{\link{classAcc}} and/or \code{\link{npelVIMP}}.}
}
\value{
a the class that \code{\link{modelAccs}} or \code{\link{modelValid}} needs.
}
\description{
These functions are intended for internal use---as such no defaults are provided and there is little to no internal parameter
checking. Using this function it is possible, however, to do things like recover VIMP data using a validating dataset, which is not
possible using the default package functions.
}
