% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwTile.R
\name{writeTiles}
\alias{writeTiles}
\title{Write multiple output tiles}
\usage{
writeTiles(inRdata, models, base, path = "./", extension = ".tif",
  echo = T, ...)
}
\arguments{
\item{inRdata}{is the image from which to generate the rendering, which can be read in using \code{\link{readTile}}}

\item{models}{a list of model names, the standard output from \code{\link{generateModels}}}

\item{base}{the base portion of the filename}

\item{path}{(optional) the path of the folder; defaults to './' the current directory}

\item{extension}{(optional) the latter part of the filename; defaults to '.tif'}

\item{echo}{(optional) should the function report its progress; defaults to TRUE}

\item{...}{other parameter(s) to pass on to \code{\link{writeTile}}, e.g. threshold, labels, et.c}
}
\value{
invisibly returns the time it took each model output to execute
}
\description{
A utility function to output all the possible model renderings model list. It builds the output filename as follows: <path><base><model class><ext>
}
\details{
Sometimes it is desirable to render all the models in a block (a list of models), for example for comparison---this function performs
that task. It is a wrapper that loops through all models and calls \code{\link{writeTile}} with a generated filename to which to
render.
}
\section{Warning}{

Runtimes may be long! On an 2015 iMac (Intel I5 quad-core 3.3 GHz, 8GB RAM, and SSD) it takes 12-15 hours to do a full rendering of a
single tile approximately 15k x 15k pixels (=225 Mpix). Also, output file sizes can be large(ish)---~2+ GB... and this is only for a
single tile. It is recommended that writeTiles be reserved for outputing comparison tiles of a smaller size, or subsampled data.
}
\examples{
# See the examples at writeTile
}
\seealso{
\code{\link{writeTile}} for the engine that does the rendering
}

